+++ src/App.tsx
@@ -18,6 +18,7 @@ import { AddBrowserModal } from './components/modals/AddBrowserModal'
 import { AddContentModal } from './components/modals/AddContentModal'
 import { AiUsageModal } from './components/modals/AiUsageModal'
 import { AuditModal } from './components/modals/AuditModal'
+import { FsBrowserModal } from './components/modals/FsBrowserModal'
 import { EditGroupModal } from './components/modals/EditGroupModal'
 import { EditProjectModal } from './components/modals/EditProjectModal'
 import { FindJumpModal } from './components/modals/FindJumpModal'
@@ -732,6 +733,7 @@ export default function App() {
         <McpIntroModal />
         <RemoteControlModal />
         <AuditModal />
+        <FsBrowserModal />
       </ErrorBoundary>
       <InAppNotifications />
       {activeView === 'agentCanvas' ? <TokenHud /> : null}